requirements.txt
scikit-learn>=1.0.0
numpy>=1.21.0
networkx>=2.6.0
matplotlib>=3.4.0
pyyaml>=5.4.0
requests>=2.26.0
